+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 19.1.23
+
+- Fix: icon positioning in password input field
+- Add possibility to hide selected option indicator on toggle
+
 ## 19.1.22
 
 - Fix: alignment bug with readonly and editable checkboxes